requirements.txt
nose
sphinx
pyyaml
